Refactor: move tag scheme merging to separate module
